Change return type of unFavoriteMovie function (from Long to Int)

+++ Data/src/main/java/ir/omidtaheri/data/repository/FavoriteMovieRepository.kt
@@ -24,7 +24,7 @@ class FavoriteMovieRepository @Inject constructor(
         )
     }
 
-    override suspend fun unFavoriteMovie(movie: FavoritedMovieDomainEntity): Long {
+    override suspend fun unFavoriteMovie(movie: FavoritedMovieDomainEntity): Int {
         return movieLocalDataSource.unFavoriteMovie(
             favoritedMovieEntityDomainDataMapper.mapToDataEntity(
                 movie
